公司

+++ app/src/main/jni/com_dy_dwvm_mt_MTLib.h
@@ -51,7 +51,7 @@ JNIEXPORT jboolean JNICALL Java_com_dy_dwvm_1mt_MTLib_setDeviceName
  * Signature: ()Z
  */
 JNIEXPORT jboolean JNICALL Java_com_dy_dwvm_1mt_MTLib_resetDeviceID
-        (JNIEnv *env, jobject, jlong deviceID);
+        (JNIEnv *env, jobject, jint deviceID);
 /*
  * Class:     com_dy_dwvm_mt_MTLib
  * Method:    dataEncrypt
